+++ tests/testthat/test-splice.R
@@ -1,0 +1,5 @@
+context('Splice')
+
+test_that('Error with empty directory', {
+  expect_error(splice("empty"), "[.]rds files")
+})